code cleanup, and moved models imports inside functions to avoid circular import
