+++ backend/api/utils/__init__.py
@@ -1,6 +1,6 @@
-from .validators import *
-from .permissions import *
+from .utils import *
 from .mixins import *
+from .permissions import *
 from .emails import *
 from .auth_backends import *
 from .custom_exception_handlers import *
