+++ backend/api/utils/mixins.py
@@ -1,41 +1,5 @@
-from django.utils.http import urlsafe_base64_decode
-from django.utils.encoding import  force_str
-from django.contrib.auth.tokens import default_token_generator
 from django.contrib.auth.password_validation import validate_password
 from rest_framework import serializers
-from ..models import (
-    User, 
-    Client,
-    Barber,
-    Admin,
-    Appointment,
-    Availability,
-    Service,
-    Review,
-    AppointmentStatus,
-)
-
-
-def get_user_from_uid_token(uidb64, token, role=None):
-    """
-    Utility function that checks if a token previously registered to a user is valid.
-    Raises serializers.ValidationError if invalid.
-    """
-    try:
-        uid = force_str(urlsafe_base64_decode(uidb64))
-
-        if role:
-            user = User.objects.get(pk=uid, role=role)
-        else:
-            user = User.objects.get(pk=uid)
-
-    except (TypeError, ValueError, OverflowError, User.DoesNotExist):
-        raise serializers.ValidationError("Invalid link.")
-    
-    if not default_token_generator.check_token(user, token):
-        raise serializers.ValidationError("Invalid or expired token.")
-
-    return user
 
 
 class PasswordValidationMixin:
@@ -52,8 +16,9 @@ class EmailValidationMixin:
     Utility mixin to handle common email validation checks
     """
     def validate_email_unique(self, attrs, user_instance=None):
-        email = attrs['email']
+        from ..models import User
 
+        email = attrs['email']
         user = User.objects.filter(email=email)
 
         if user_instance:
@@ -71,8 +36,9 @@ class UsernameValidationMixin:
     Utility mixin to handle common username validation checks
     """
     def validate_username_unique(self, attrs, user_instance=None): 
-        username = attrs['username']
+        from ..models import User
 
+        username = attrs['username']
         user = User.objects.filter(username=username)
 
         if user_instance:
@@ -90,6 +56,8 @@ class UIDTokenValidationSerializer(serializers.Serializer):
     Utility serializer that handlles token checks, from which other serializers inherit
     """
     def validate_uid_token(self):
+        from .utils import get_user_from_uid_token
+
         uidb64 = self.context.get('uidb64')
         token = self.context.get('token')
 
@@ -110,6 +78,8 @@ class UserValidationMixin:
     Mixin to validate that a user_id from context exists and is active. Also adds 'user' to attrs.
     """
     def validate_user(self, attrs):
+        from ..models import User
+
         user_id = self.context.get('user_id')
 
         try:
@@ -126,6 +96,8 @@ class AdminValidationMixin:
     Mixin to validate that a admin_id from context exists and is active. Also adds 'admin' to attrs.
     """
     def validate_admin(self, attrs):
+        from ..models import Admin
+
         admin_id = self.context.get('admin_id')
 
         try:
@@ -142,6 +114,8 @@ class ClientValidationMixin:
     Mixin to validate that a client_id from context exists and is active. Also adds 'client' to attrs.
     """
     def validate_client(self, attrs):
+        from ..models import Client
+
         client_id = self.context.get('client_id')
 
         try:
@@ -158,6 +132,8 @@ class BarberValidationMixin:
     Mixin to validate that a barber_id from context exists and is active. Also adds 'barber' to attrs.
     """
     def validate_barber(self, attrs):
+        from ..models import Barber
+
         barber_id = self.context.get('barber_id')
 
         try:
@@ -189,6 +165,8 @@ class AppointmentValidationMixin:
         return attrs
 
     def validate_appointment_date_and_slot(self, attrs):
+        from ..models import Appointment, Availability, AppointmentStatus
+
         client = attrs['client']
         barber = attrs['barber']
         appointment_date = attrs['date']
@@ -216,6 +194,8 @@ class AppointmentValidationMixin:
         return attrs
     
     def validate_find_appointment(self, attrs):
+        from ..models import Appointment, AppointmentStatus
+
         client = attrs['client']
         appointment_id = self.context.get('appointment_id')
 
@@ -239,6 +219,8 @@ class AvailabilityValidationMixin:
     - Validates the existence of an availability entry for the given barber and specified ID before allowing retrieval or update operations.
     """
     def validate_availability_date(self, attrs, availability_instance=None):
+        from ..models import Availability
+
         barber = attrs['barber']
         availability_date = attrs['date']
 
@@ -253,6 +235,8 @@ class AvailabilityValidationMixin:
         return attrs
 
     def validate_find_availability(self, attrs):
+        from ..models import Availability
+        
         barber = attrs['barber']
         availability_id = self.context.get('availability_id')
 
@@ -273,6 +257,8 @@ class ServiceValidationMixin:
     - Ensures a service with the given ID exists and is owned by the specified barber before proceeding with actions that need to fetch or validate a particular service.
     """
     def validate_service_name(self, attrs, service_instance=None):
+        from ..models import Service
+
         barber = attrs['barber']
         service_name = attrs['name']
 
@@ -287,6 +273,8 @@ class ServiceValidationMixin:
         return attrs
 
     def validate_find_service(self, attrs):
+        from ..models import Service
+
         barber = attrs['barber']
         service_id = self.context.get('service_id')
 
@@ -307,6 +295,8 @@ class ReviewValidationMixin:
     - Ensures a review exists and belongs to the requesting client when retrieving or modifying a review.
     """
     def validate_appointment_for_review(self, attrs):
+        from ..models import Appointment, Review, AppointmentStatus
+
         client = attrs['client']
         appointment_id = self.context.get('appointment_id')
 
@@ -328,6 +318,8 @@ class ReviewValidationMixin:
         return attrs
     
     def validate_find_review(self, attrs):
+        from ..models import Review, AppointmentStatus
+
         client = attrs['client']
         review_id = self.context.get('review_id')
 
@@ -348,6 +340,7 @@ class GetAdminsMixin:
         """
         Returns Admin queryset in the system.
         """
+        from ..models import Admin
         return Admin.objects.all()
     
     def get_admin_private(self, admin):
@@ -374,6 +367,7 @@ class GetBarbersMixin:
         Returns Barber queryset in the system.
         If show_all is True, returns all barbers.
         """
+        from ..models import Barber
         return Barber.objects.filter(is_active=True) if not show_all else Barber.objects.all()
     
     def get_barber_public(self, barber):
@@ -415,6 +409,7 @@ class GetClientsMixin:
         Returns Client queryset in the system.
         If show_all is True, returns all clients.
         """
+        from ..models import Client
         return Client.objects.filter(is_active=True) if not show_all else Client.objects.all()
     
     def get_client_public(self, client):
@@ -454,6 +449,7 @@ class GetAvailabilitiesMixin:
         Returns Availability queryset for a specific barber.
         If show_all is True, returns all availabilities.
         """
+        from ..models import Availability
         return Availability.objects.filter(barber_id=barber_id) if not show_all else Availability.objects.all()
     
     def get_availability_public(self, availability):
@@ -478,6 +474,7 @@ class GetServicesMixin:
         Returns Service queryset for a specific barber.
         If show_all is True, returns all services.
         """
+        from ..models import Service
         return Service.objects.filter(barber_id=barber_id) if not show_all else Service.objects.all()
     
     def get_service_public(self, service):
@@ -502,6 +499,8 @@ class GetAppointmentsMixin:
         Returns Appointment queryset filtered by barber or client.
         If show_all is True, returns all appointments.
         """
+        from ..models import Appointment
+
         if show_all:
             return Appointment.objects.all()
         
@@ -539,6 +538,8 @@ class GetReviewsMixin:
         Returns Review queryset filtered by barber or client.
         If show_all is True, returns all reviews.
         """
+        from ..models import Review
+
         if show_all:
             return Review.objects.all()
