+++ backend/api/utils/permissions.py
@@ -1,5 +1,4 @@
 from rest_framework.permissions import BasePermission
-from ..models import Roles 
 
 
 class IsAdminRole(BasePermission):
@@ -7,6 +6,8 @@ class IsAdminRole(BasePermission):
     Allows access only to users with the ADMIN role.
     """
     def has_permission(self, request, view):
+        from ..models import Roles 
+
         return (
             request.user 
             and request.user.is_authenticated 
@@ -19,8 +20,11 @@ class IsClientRole(BasePermission):
     Allows access only to users with the CLIENT role.
     """
     def has_permission(self, request, view):
+        from ..models import Roles 
+
         return (
-            request.user and request.user.is_authenticated
+            request.user 
+            and request.user.is_authenticated
             and request.user.role == Roles.CLIENT.value
         )
 
@@ -30,7 +34,10 @@ class IsBarberRole(BasePermission):
     Allows access only to users with the BARBER role.
     """
     def has_permission(self, request, view):
+        from ..models import Roles 
+        
         return (
-            request.user and request.user.is_authenticated
+            request.user 
+            and request.user.is_authenticated
             and request.user.role == Roles.BARBER.value
         )
