+++ backend/manage.py
@@ -1,12 +1,10 @@
 #!/usr/bin/env python 
-"""Django's command-line utility for administrative tasks."""
-import os
 import sys
 
 
 def main():
     """
-    Run administrative tasks.
+    jango's command-line utility for administrative tasks.
     """
     try:
         from django.core.management import execute_from_command_line
